convert AtomicU64 into AtomicUsize and u64s into usizes

+++ src/tests.rs
@@ -8,15 +8,15 @@ fn test_reachability() {
 
 #[test]
 fn test_download() {
-    let total_bytes_counter = Arc::new(AtomicU64::new(0));
-    let current_down = Arc::new(AtomicU64::new(0));
+    let total_bytes_counter = Arc::new(AtomicUsize::new(0));
+    let current_down = Arc::new(AtomicUsize::new(0));
     let exit_signal = Arc::new(AtomicBool::new(false));
     download_test(1, &total_bytes_counter, &current_down, &exit_signal).unwrap();
 }
 
 #[test]
 fn test_upload() {
-    let upload_counter = Arc::new(AtomicU64::new(0));
+    let upload_counter = Arc::new(AtomicUsize::new(0));
     let exit_signal = Arc::new(AtomicBool::new(false));
     upload_test(1, &upload_counter, &exit_signal).unwrap();
 }
